bidder/requirements.txt
flask
redis
kafka-python
numpy